Sync the last 10 months of API docs
This repository is a mirror of a subdirectory of an internal Slack repository.
Due to some changes in internal tooling our syncing scripts broke, and we
stopped updating this mirror.

While we were rebuilding those scripts we realized that syncing every
individual commit into this repository is going to be too much work for us,
and that most of the value of this mirror comes from making the JSON files
available programatically, not from seeing the details of every individual
change.

From today we're planning on doing periodic commits every few weeks containing
everything that's changed since the last sync, along with occasional special
commits around specific feature launches. We hope this is frequent enough for
everyone.

Sorry this repository went unmaintained for so long! :heart: